Share CFC structured result primitives
