+++ packages/runner/src/builtins/llm-dialog.ts
@@ -47,10 +47,16 @@ import {
 } from "../query-result-proxy.ts";
 import { ContextualFlowControl } from "../cfc.ts";
 import { type CfcLabelView, cfcLabelViewForCell } from "../cfc/label-view.ts";
-import { cfcLabelPathPrefixMatches } from "../cfc/label-view-core.ts";
+import {
+  cfcConfidentialityForObservationNode,
+  cfcObservationFitsCeiling,
+  type CfcObservationResult,
+  joinCfcObservedConfidentiality,
+  uniqueCfcAtoms,
+} from "../cfc/observation.ts";
+import { cfcSchemaToObject, resolveCfcSchemaRefs } from "../cfc/schema-refs.ts";
 import { createFrozenRequestSnapshot } from "../cfc/request-snapshot.ts";
 import { enqueueSinkRequestPostCommitEffect } from "../cfc/sink-request.ts";
-import { deepEqual } from "@commonfabric/utils/deep-equal";
 import { resolveLink } from "../link-resolution.ts";
 import { internalVerifierRead } from "../storage/reactivity-log.ts";
 import type { RawBuiltinResult } from "../module.ts";
@@ -98,10 +104,7 @@ function stripInjectedResult(
 // --------------------
 
 type ToolKind = "handler" | "cell" | "pattern";
-type LLMObservationSerializationResult = {
-  value: unknown;
-  observedConfidentiality: readonly unknown[];
-};
+type LLMObservationSerializationResult = CfcObservationResult;
 
 type SerializeForLLMObservationParams = {
   value: unknown;
@@ -151,7 +154,7 @@ function resolveRefsForLLM(
   const toObj = (s: unknown) =>
     s === false
       ? ({ type: "object", properties: {} } as Record<string, unknown>)
-      : ContextualFlowControl.toSchemaObj(
+      : cfcSchemaToObject(
         typeof s === "boolean" ? s : (s as JSONSchema) ?? undefined,
       );
 
@@ -171,7 +174,7 @@ function resolveRefsForLLM(
         // Circular or too deep — truncate
         return { type: "object", additionalProperties: true };
       }
-      const resolved = ContextualFlowControl.resolveSchemaRefs(
+      const resolved = resolveCfcSchemaRefs(
         nodeObj,
         schema,
       );
@@ -403,59 +406,6 @@ function simplifySchemaForContext(
   return simplified as JSONSchema;
 }
 
-function joinObservedConfidentiality(
-  parts: Iterable<readonly unknown[] | undefined>,
-): readonly unknown[] {
-  const joined: unknown[] = [];
-  for (const part of parts) {
-    if (!Array.isArray(part)) {
-      continue;
-    }
-    joined.push(...part);
-  }
-  return ContextualFlowControl.uniqueAtoms(joined);
-}
-
-function confidentialityForCurrentNode(
-  schema: JSONSchema | undefined,
-  labelView: CfcLabelView | undefined,
-  logicalPath: readonly string[],
-): readonly unknown[] {
-  const joined: unknown[] = [];
-
-  if (isRecord(schema) && isRecord(schema.ifc)) {
-    joined.push(...(schema.ifc.confidentiality ?? []));
-  }
-
-  if (labelView !== undefined) {
-    for (const entry of labelView.entries) {
-      if (!cfcLabelPathPrefixMatches(entry.path, logicalPath)) {
-        continue;
-      }
-      joined.push(...(entry.label.confidentiality ?? []));
-    }
-  }
-
-  return ContextualFlowControl.uniqueAtoms(joined);
-}
-
-function fitsObservationCeiling(
-  confidentiality: readonly unknown[],
-  observationMaxConfidentiality: readonly unknown[] | undefined,
-): boolean {
-  if (
-    observationMaxConfidentiality === undefined ||
-    observationMaxConfidentiality.length === 0 ||
-    confidentiality.length === 0
-  ) {
-    return true;
-  }
-
-  return confidentiality.every((value) =>
-    observationMaxConfidentiality.some((allowed) => deepEqual(allowed, value))
-  );
-}
-
 function observationLinkForValue(
   value: unknown,
   logicalPath: readonly string[],
@@ -511,13 +461,13 @@ function serializeForLLMObservation(
     };
   }
 
-  const nodeConfidentiality = confidentialityForCurrentNode(
+  const nodeConfidentiality = cfcConfidentialityForObservationNode({
     schema,
     labelView,
     logicalPath,
-  );
+  });
   if (
-    !fitsObservationCeiling(
+    !cfcObservationFitsCeiling(
       nodeConfidentiality,
       observationMaxConfidentiality,
     )
@@ -634,7 +584,7 @@ function serializeForLLMObservation(
 
     return {
       value: serialized,
-      observedConfidentiality: joinObservedConfidentiality(observedParts),
+      observedConfidentiality: joinCfcObservedConfidentiality(observedParts),
     };
   }
 
@@ -665,7 +615,7 @@ function serializeForLLMObservation(
 
   return {
     value: serialized,
-    observedConfidentiality: joinObservedConfidentiality(observedParts),
+    observedConfidentiality: joinCfcObservedConfidentiality(observedParts),
   };
 }
 
@@ -1556,7 +1506,7 @@ function buildAvailableCellsDocumentationWithObservation(
   const entries = [...seenPaths.values()].map((v) => v.entry);
   return {
     docs: "\n\n# Available Cells\n\n" + entries.join("\n\n"),
-    observedConfidentiality: joinObservedConfidentiality(
+    observedConfidentiality: joinCfcObservedConfidentiality(
       [...seenPaths.values()].map((value) => value.observedConfidentiality),
     ),
   };
@@ -1587,17 +1537,16 @@ function getObservedDialogMessages(
   observedConfidentiality: readonly unknown[];
 } {
   const labelView = cfcLabelViewForCell(messagesCell);
-  const observedConfidentiality = joinObservedConfidentiality(
+  const observedConfidentiality = joinCfcObservedConfidentiality(
     messages.map((_message, index) => {
       const stored = messageObservations[index.toString()];
       if (Array.isArray(stored)) {
         return stored;
       }
-      return confidentialityForCurrentNode(
-        undefined,
+      return cfcConfidentialityForObservationNode({
         labelView,
-        [index.toString()],
-      );
+        logicalPath: [index.toString()],
+      });
     }),
   );
 
@@ -1615,7 +1564,7 @@ function mergeDialogMessageObservations(
   };
   for (const update of updates) {
     const key = update.index.toString();
-    merged[key] = ContextualFlowControl.uniqueAtoms([
+    merged[key] = uniqueCfcAtoms([
       ...(merged[key] ?? []),
       ...(update.observedConfidentiality ?? []),
     ]) as unknown[];
@@ -1915,9 +1864,7 @@ function toolAllowsObservedConfidentiality(
     return true;
   }
 
-  return observedConfidentiality.every((value) =>
-    maxConfidentiality.some((allowed) => deepEqual(allowed, value))
-  );
+  return cfcObservationFitsCeiling(observedConfidentiality, maxConfidentiality);
 }
 
 async function executeToolCalls(
@@ -2994,7 +2941,7 @@ Some operations (especially \`invoke()\` with patterns) create "Pages" - running
       .get() as DialogMessageObservationMap) ??
       {},
   );
-  const requestObservedConfidentiality = joinObservedConfidentiality([
+  const requestObservedConfidentiality = joinCfcObservedConfidentiality([
     visibleMessages.observedConfidentiality,
     cellsDocs.observedConfidentiality,
   ]);
